Make setup more flexible
Currently there's no way to have an up-to-date `envdir` in a service
and ensure it's down at the same time. This is useful for running
management commands with proper environment.

This change will allow us for example in webshell to have
`$service_ensure = 'stopped'` and `$service_status = 'enabled'`
so we can have an up-to-date envdir and at the same time not run
Gunicorn process. This would be done in `management` hosts where we
don't need a running WSGI server, but we need current envdir to run
management commands (like migrations).

Also I believe most of the comments in this module were misleading (at least)
and the behaviour when `$service_ensure != 'running'` was wrong.
If I understand the meaning of `down` file correctly from
http://cr.yp.to/daemontools/supervise.html
we want it to be "present" if we don't want to start the service
and to be "absent" to operate normally (i.e. run the service).

As for the backwards compatibility, we either use a `running`-`running`
combination (default values for `daemontools::setup`) or
`disabled`-`stopped` when we want to put down a service
(see `internal/laterpay/manifests/apps/webshell/{worker,beat}.pp`),
so we should be fine.